requirements.txt
async-timeout==5.0.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-documentai==3.5.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
heyoo==0.1.2
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.0
redis==6.0.0
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.21.0
